Issue: MCUX-56506 [RT1180] Build Warning on WIFI Submodule
[Issue]
Build warnings observed in different compilers.

[Fix]
Fixed build warnings raised by Dapeng build server for various compilers.
Ran and applied file formatting scripts.

[Validation done]
Build wifi_cli application successfully with armgcc, mdk and IAR compilers.

[Open Issues/Todo]
None.

[Reviewers]
Abhinav Kulkarni, Sandip Dalvi, Supriya Inamdar, Sanjay Yadav.

+++ wlcmgr/wlan_enhanced_tests.c
@@ -564,7 +564,7 @@ static void print_ds_rate(wlan_ds_rate ds_rate)
                 {
                     (void)PRINTF("    MCS:  Auto\n");
                 }
-                (void)PRINTF("    Rate: %f Mbps\r\n", (float)datarate->rx_data_rate / (float)2);
+                (void)PRINTF("    Rate: %.2f Mbps\r\n", (double)datarate->rx_data_rate / 2);
             }
         }
 #endif
